Change hazelcast.xml default to tcp-ip + localhost
Hazelcast communication using multi-cast with default address may
interfere with other host who happens to be on the same network.
e.g.,When the code is checked-out and user immediately run mvn test, etc.

This is a short term workaround to address the issue by
changing the default to tcp-ip and forming cluster using localhost only.

Change-Id: I762edd6b37ccdb689c8b6493b60fdca7a8f1a0e1