Moved the initialization of the ArpCache to ProxyArpManager's startUp method.
We have a timing issue with the initialization of the hazelcast instances when
running hazelcast as the datastore. The instance is started by the notificaiton
framwork relatively late (during the startUp method). The datastore needs to
connect to this instance and can be initialized earlier with unlucky timing.

This is a workaround to make the ArpCache initialize its datastore during the
startUp method, so hopefully this happens after the notification framework's
startUp method has been called.

Change-Id: Ie3bdc473630c64bfa912688747988a31db3fee72