Add a utility class checker for unit tests
Added a utility class checker and a unit test for it.  The
utility class checker makes sure that a utility class adheres
to the following rules:

    - the class must be declared final
    - the class must have only one constructor
    - the constructor must be private and inaccessible to callers
    - the class must have only static methods

Putting this checker in your unit tests will prevent future modifications
to your class that break the utility class assumptions.  The checker
also triggers an object creation through reflection that will
cause coverage events from the test.

Change-Id: Ia469908aefd5ad5c9a7387e6168576d50da01b3e