Re-enabled arp request expiry and remove the device deletion mechansim.
We now remove arp requests from our data structures if the host doesn't
respond within a certain amount of time.

Also, we used to delete the device if it didn't respond to an ARP request in
an effort to remove stale data. Now we have a proper device cleanup thread
for this.

Change-Id: Idd50cc89d184fac12bbcc0fc6726746e1e1ca693