Removed the ipAddress and activeSince fields from OnosDevice.
We decided that the Device class would be a simple mapping between attachment port and MAC address. IP information is stored separately so there is no more use for the ipAddress field.

I don't know what the activeSince field was doing. We never use it - we only use the lastSeenTimestamp timestamp.

Change-Id: I252d564358f619d7b26fcbeb2ad9c8bc274f864d