Create a config option to tell ONOS to only send OF1.0 HELLO packets.
Switches and controllers should negotiate down to the highest OpenFlow version
mutually supported, however our testing with the FlowSpace Firewall shows that
the firewall 'switch' cannot correctly parse OF1.3 HELLO packets because they
are longer than OF1.0 HELLO packets. This seems to be a bug in the switch
but we can work around it in ONOS for now.

The option can be enabled by adding the following to onos.properties:
 net.floodlightcontroller.core.FloodlightProvider.useOnly10 = true

Change-Id: I51e5cc4f64124b6a90e8b91fc30d286fbdaf6915