Getter REST APIs for Intents, FlowManager, and MatchAction.
Resource and Routable classes are implemented but will need to be registered
with the REST server when the module implementations are in place.

Right now they use default JSON serializers for everything apart from the ID
classes, which couldn't be default serialized. We can write custom serializers
for the objects if the need arises.

ONOS-1878
ONOS-1876
ONOS-1877

Change-Id: I48df3532afcf1e8bec104a52d58e23a6af4cffbe