Bug fix: Topology WebSocket timeout if there are no events to transmit
Apparently, if there are no topology-related events to transmit to
the client, the WebSocket will automatically timeout after several minutes.

The solution is to use the WebSocket embedded PING-PONG mechanism.
Implementation-wise, if there are no events to transmit for 30 seconds,
ONOS will transmit a PING message. If three PING messages are transmitted,
and there is no PONG, then the WebSocket is closed. The total
timeout time is approximately 30x(3+1) = 120 seconds.
Note that the PONG message is automatically transmitted by the Web browsers,
so no code change is required to the ONOS GUI.

This fixes bug ONOS-1932

Also, updated websocket-api artifact in pom.xml from 1.0 to 1.1

Change-Id: I474e04e81a6f613774bb987f17a6a0b5d3afa99d