Fix a bug in the Link Discovery module when a switch is disconnected.
Apparently, if the switch is disconnected from the controller and
then reconnected, the discovered link state in the LinkDiscoveryManager
is not cleaned-up properly.
As a result, the links are not published when the switch reconnects.

Depends on timing, the links might actually be removed from the
Topology Manager, but won't be added back.

Change-Id: I2c1e1da9190047a93bd917c79bac9babeeb4ef6f